Changed the test prints to be less cluttered and more helpful, in the process changed the Cucumber package structure and refactored the ServerRule class to be an anonymous class within RunCukesTest
