--- src/test/java/pingis/ServerRule.java
@@ -1,25 +1,0 @@
-package pingis;
-
-import org.junit.rules.ExternalResource;
-import org.springframework.boot.SpringApplication;
-import org.springframework.context.ConfigurableApplicationContext;
-
-class ServerRule extends ExternalResource{
-
-    private final int port;
-    ConfigurableApplicationContext app;
-
-    public ServerRule(int port) {
-        this.port = port;
-    }
-
-    @Override
-    protected void before() throws Throwable {
-        this.app = SpringApplication.run(Application.class);
-    }
-
-    @Override
-    protected void after() {
-        app.close();
-    }
-}
