+++ src/test/java/pingis/cucumber/Stepdefs.java
@@ -1,4 +1,4 @@
-package pingis;
+package pingis.cucumber;
 
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
